requirements.txt
pymysql-1.0.2
Flask-2.0.2